Former-commit-id: 4723fbcc3895b8ecf3f8e91f1c2ab8272d43bdc4

+++ Sources/Internal/DLC/Patcher/PatchFile.cpp
@@ -116,7 +116,7 @@ bool PatchInfo::ReadString(File* file, String &str)
 {
     bool ret = false;
     uint32 len = 0;
-    uint32 rlen = file->Read(&len);
+    size_t rlen = file->Read(&len);
 
     if(rlen == sizeof(len))
     {
@@ -124,7 +124,7 @@ bool PatchInfo::ReadString(File* file, String &str)
         {
             String tmpStr;
             tmpStr.resize(len);
-            rlen = file->Read(&tmpStr[0], tmpStr.size());
+            rlen = file->Read(&tmpStr[0], static_cast<uint32>(tmpStr.size()));
 
             if(rlen == tmpStr.size())
             {
